docs(user-journey): add user journey map based on my understanding of requirements
parts of these may be descoped for future since we want to iterate fast and i don't want to be 'gold plating' my app so much that i'm not making progress. but, this is good to document regardless as my understanding of the requirements in text format on odin project in terms of how a user may want to be interacting with any app that i build for this inventory management purpose. btw, we're allow for _null_ categories