fix(populatedb): escape single-quotation characters
identified that the reason why i was getting errors was because i didn't escape the single quotation character when inserting rows into the items table. postgresql would have probably thought that was the end of the string i was trying to insert. i tried to use double-quotation but then got error that 'column does not exist', so just backslashed to escape.