feat(populatedb): pre-populate local database with npm script (#2)
* docs(tables): add diagram of designed tables for v1

it does look a bit simple, but i want to get an mvp out first. this diagram will form the basis of this branch/PR

* feat(db): use local (localhost) db URI for both pg.Pool and pg.Client

not sure where I'm going to be deploying the code (and thus hosting the PostgreSQL database) yet, so will just work on local DB. use dotenv (.env file) to hold local db's uri - db is called 'coffee_inventory_db' by the way. add script for both creating and populating tables in populatedb.js. these reflect the wireframed fake data to illustrate logic and test it manually when I run the script

* fix(package.json): update seed-db npm script to correct path

forgot to update this when i moved all the code files into src/

* fix(populatedb): add semicolon to end of POPULATE_TABLES_QUERY to prevent error

this was manually tested by creating in psql in terminal, and it did work

* docs(README): add instruction (to be refined) that the database needs to be created locally first before the tables can be inserted

perhaps there's a way to do this through the populatedb script, but i wasn't able to run 'CREATE DATABASE IF NOT EXISTS..' before the script begins... might be for a future enhancement of this script? not sure if it's even necessary for prod db though...

* fix(populatedb): escape single-quotation characters

identified that the reason why i was getting errors was because i didn't escape the single quotation character when inserting rows into the items table. postgresql would have probably thought that was the end of the string i was trying to insert. i tried to use double-quotation but then got error that 'column does not exist', so just backslashed to escape.

* docs(readme): add instructions around .env file

this will make it easier for anyone to see local postgresql db without having to dig through the code (specifically, the populatedb script). just better dx for any future devs looking at this project!

* fix(populatedb): properly escape single quotation marks. increase length of description of coffee beans

escaped using double single-quotation marks like this: https://stackoverflow.com/a/38701038 . increased length of the description - 280 characters was certainly not enough and it's just better to just future proof this i think!

* refactor: move name of db to constants file. refactor(pool.js): align the creation of pool (not really needed for this PR) with shape of .env file

keeps code DRY by importing from the constants. might need other constants (not sure yet) but would be good to have this to future-proof.